list, detail에 작성자 정보가 나오도록

+++ src/main/webapp/jsp/article/detail.jsp
@@ -8,6 +8,10 @@
 <%
 	Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
 %>
+<%
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
+%>
+
 
 <html>
 <head>
@@ -15,19 +19,29 @@
 <title>게시물 상세보기</title>
 </head>
 <body>
+
+
+
+
 	<h1><%= (int) articleRow.get("id") %>번 게시물</h1>
 	
 	<div>번호 : <%= (int) articleRow.get("id") %></div>
 	<div>날짜 : <%= (LocalDateTime) articleRow.get("regDate") %></div>
+	<div>작성자 : <%= (String) articleRow.get("writerName") %></div>
 	<div>제목 : <%= (String) articleRow.get("title") %></div>
 	<div>내용 : <%= (String) articleRow.get("body") %></div>
 	<div>
-<!-- 		onclick 은 앞의 href보다 먼저 실행된다 -->
-<!-- return false;일경우 기능이 되지 않는다 -->
-<!-- confirm역시 알림 창의 일종이다-->
-		<a href="list">목록</a>
+	
+		<a href="list">목록</a>	
+	<%
+	if (loginedMemberId != -1) {
+	%>
+	
 		<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
 		<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+	<%
+	}
+	%>
 	</div>
 </body>
 </html>
